test(harness): declare v_port_* control seam in the stub port.h
The stub tests/stubs/port.h declared only the v_port_hw_* facade, so every
kernel TU that calls v_port_trigger_pendsv / v_port_halt / v_port_get_psp /
v_port_disable_interrupts / v_port_mpu_init / v_port_hw_cpu_idle /
v_port_hw_active_irq_priority tripped -Wimplicit-function-declaration (the
symbols are stubbed in .c but were undeclared). Adding the two SVC/task-heap
test binaries multiplied the count. Declare them (matching the real port.h) —
drops the host build from 19 warnings to 4.

Also fix a 64-bit-host pointer-narrowing warning in utils.c's stack-overflow
log guard: cast mem_block through uintptr_t before the uint32_t compare.

Remaining 4 warnings are pre-existing in test_structure.c (TEST_ASSERT's
return in a non-void helper) and test_scheduler.c (Task_Status sign-compare),
untouched here.

+++ kernel/utils.c
@@ -701,12 +701,12 @@ void v_log(Log_Type type, const char *msg, ...) {
   if (scheduler_running && current_task && current_task->magic == TCB_MAGIC) {
     uint32_t psp = v_port_get_psp();
     // 300 bytes is safe for most panics, but let's be even safer with 320
-    if (psp != 0 && psp < (uint32_t)current_task->mem_block + 320) {
+    if (psp != 0 && psp < (uint32_t)(uintptr_t)current_task->mem_block + 320) {
       v_panic(__FILE__, __LINE__,
               "Stack overflow detected in task %u during log! SP: 0x%x, Limit: "
               "0x%x",
               (unsigned)current_task->task_id, (unsigned)psp,
-              (unsigned)((uint32_t)current_task->mem_block + 320));
+              (unsigned)((uint32_t)(uintptr_t)current_task->mem_block + 320));
     }
   }
